Harden clipboard, addon trust, NLP, theme, and IME lifecycle paths
Deep audit pass over :app and :lib:*. Highlights:

Security
- clipboard: never clone EXTRA_IS_SENSITIVE media to on-disk storage
- addon: refuse multi-signer/rotated-cert signing pins (bypass surface)
- IME: extend FLAG_SECURE to user-toggled incognito fields
- dict import: scrub the decrypted plaintext sniff buffer

Correctness
- bidi: UAX#9 L2 visual reorder + code-point-safe RTL reversal
- clipboard provider: return the orientation cursor; select the real column
- subtype switch: index-based wrap instead of the DEFAULT sentinel
- KLC metadata: slice by token length, not the uppercased token
- snygg: deserialize fractional-dp cut-corner and clamp overflowing max-lines
- translate: preserve inter-sentence whitespace when stitching pieces
- voice: snapshot/restore the transcript buffer on editor-edit failure
- dict-import undo: key-match inserted rows so concurrent learns survive

Concurrency / leaks
- n-gram stores + media-provider cache: ConcurrentHashMap / AtomicInteger
- clipboard close() cancels the whole IO scope; clone ids no longer collide
- IME onDestroy clears inline-autofill views; backup/restore close on dispose
- crash handler guards its stacktrace read; KMP parser caps entries/size

UX
- confirm-gate 'Reset all learning'; show true per-tool MCP switch state
- voice IME switch surfaces the picker instead of a silent no-op

Build gates not run here (no Android SDK on this host).

+++ lib/snygg/src/main/kotlin/org/florisboard/lib/snygg/value/SnyggShapeValue.kt
@@ -139,10 +139,15 @@ data class SnyggCutCornerDpShapeValue(
         override fun deserialize(v: String) = runCatching<SnyggValue> {
             val map = snyggIdToValueMapOf()
             spec.parse(v, map)
-            val topStart = map.getInt(CornerSizeTopStart).dp
-            val topEnd = map.getInt(CornerSizeTopEnd).dp
-            val bottomEnd = map.getInt(CornerSizeBottomEnd).dp
-            val bottomStart = map.getInt(CornerSizeBottomStart).dp
+            // getFloat, not getInt: the spec declares all four corners as float(DpUnit)
+            // (matching the SnyggRoundedCornerDpShapeValue sibling), so a fractional dp
+            // such as cut-corner(4.5dp,...) parses to "4.5"; getInt would throw
+            // NumberFormatException, be swallowed by runCatching, and silently discard
+            // the whole shape on deserialize even though serialize emits the fraction.
+            val topStart = map.getFloat(CornerSizeTopStart).dp
+            val topEnd = map.getFloat(CornerSizeTopEnd).dp
+            val bottomEnd = map.getFloat(CornerSizeBottomEnd).dp
+            val bottomStart = map.getFloat(CornerSizeBottomStart).dp
             return@runCatching SnyggCutCornerDpShapeValue(topStart, topEnd, bottomEnd, bottomStart)
         }
     }
